refactor workbench sandbox routes
